Filter reset script Redis keys
